requirements.txt
aiohttp
asyncio
requests
objdict


chatterbot



peewee_async
psycopg2
aiopg
asyncio_redis


